Improved IR Robustness, PID updates on the fly...
Quite a few changes, but the big one is about to come next. The IR timer2 is now only started once a start signal has been detected and is stopped at the end of the packet. This saves loads of unnecessary interrupts.

The start bit for the IR packet is checked to see if it is valid (between 360 - 440us). If it's incorrect the receive is aborted and searching occurs 2ms later. This provides significantly improved interference rejection.

Variable packet lengths are now supported for IR from 1-32 bits. We're now using 20bits (8-YAW_SETPOINT, 8-SPEED, 4-INFO). The info bits are being used for PID tuning on the fly, and to support the next update of ESC support.

The PID algorithm is updated every 10ms rather than 5ms since there's no point going faster as the Yaw value is only updated every 10ms.