Untested 2 packets
We've optimised a few variables down in size (ie. longs to ints, ints to chars etc)
We've changed the EXPECTED_BITS from 20 to 17
After IR decoding we check whether it's a Type A or Type B Packet and update the correct globals.
We've done away with channel & PID changes for the time being, we'll get those back soon
A few comments have been altered, a few lines have been changed for clarity reasons.

Haven't altered the start bit search times, so with a shorter packet we'll be searching more than is necessary. Could increase these by 2.4ms.

This needs altered Tx software but the expectation is that this all gets pulled into the master.

This all desperately needs testing! I'm not sure the RF stuff is working atm, unsure why.